refactor: extract game logic
